Connection

The connection to the SAOImage DS9 is performed using the XPA interface. By default, all routines connect to the first DS9 window found. Optionally, however, the user can select a different window using ds9select.

All further operations will be performed by default over the connected window (i.e., the current access_point).

User interface

DS9ui.ds9 — Function

ds9([name]; method="local", path="ds9")

Launch the DS9 application.

By default the application will be named using the current PID.

The method optional keywords is used to set the XPA communication method: "local" is the recommended way for local executions.

This function authomaticall sets the default access point, so that all further requests are forwarded to the newly open DS9 window.

DS9ui.ds9select — Function

ds9select(ident="DS9:*"; method="local"; silent=false, interactive=true)

Select a DS9 window for further interactions.

If multiple DS9 windows match the ident, the user can select the correct window using a simple interface; if interactive=false the first matching window is selected.

Warning

If method is not undefined, the environment variable XPA_METHOD is set to the corresponding string. Typically the most reliable connection is local, the default value.

DS9ui.ds9close — Function

ds9close([access_point])

Close the current DS9 session.

Internal routines and variables

DS9ui.access_point — Constant

The current XPA access point

DS9ui.connect — Function

connect(ident="DS9:*"; method="local")

Establish a new connection to DS9 using the given ident.

If method is not undefined, the environment variable XPA_METHOD is set to the corresponding string. Typically the most reliable connection is local, the default value.

DS9ui._current_ap — Function

_current_ap()

The default access point.

This function just returns access_point if this variable is set, or establishes a new connection using connect.