Interface

High-level interface

DS9ui.ds9image — Function

ds9image([access_point,] image; usefile=true, wcs=nothing)
ds9image!([access_point,] image; usefile=true, wcs=nothing)

Display the image using SAOImage DS9.

In the bang (ds9image!) version, the image replaces the content of the current frame.

DS9ui.ds9delete — Function

ds9delete([access_point])

Delete the current frame of DS9.

DS9ui.ds9mask — Function

ds9mask([access_point,] name::String=""; coords=:image, silent=true, 
        full=false, usefits=false)

Return a mask of the regions defined in the DS9 window.

The optional argument name is the name of the group of regions to extract. If name is an empty string, all regions are extracted.

Keyword Arguments

ident: the identifier of the DS9 window.

coords: the type of coordinates to return: can be :image, :physical, :fk5, :galactic

silent: if true, no warning is issued when unknown regions are found.

full: if true, the mask is the full image. If false, the mask is cropped to the smallest rectangle containing all regions.

usefits: if true, the mask is extracted from the FITS file. If false, the mask is extracted from the DS9 window.

The return value is a tuple (image, mask), where image is the image extracted from the DS9 window (that is, the pixel values), and mask is a boolean OffsetArray with the mask (that is, true inside the regions and false outside).

DS9ui.ds9fitprofiles — Function

ds9fitprofiles([access_point,] name; modes=[:gaussian],
               usefits=false, silent=false, max_calls=1_000_000)

Fit a region in DS9 with elliptical profiles.

This function extract a region in DS9 and then fits the data with a superposition of profiles identified by modes. A background value for the sky is always added. Known profiles are :gaussian (the default), :moffat, :lorentzian, :sersic, :exponential, :devaucouleurs, and :king. Note that modes is an array: all given profiles will be combined in the specified order for the final fit.

Other options control the way data are retrieved from DS9 (usefits), the output to print (silent), and the maximum number of calls to the chi-square funtion used for the fit (max_calls).

The function returns the best-fit parameters (as a vector) and a matrix with the residuals (data - model).

DS9ui.ds9pickobj — Function

ds9pickobj([access_point]; ident="DS9:*", mode=:gaussian, r=32, 
           shape=:circle, silent=false, usefits=false)

Fits interactively a region in DS9 with an elliptical profile.

This function allows the selection of a circular or quadratic region in DS9 and fits it with a given profile using ds9fitprofiles.

Low-level interface

DS9ui.ds9get — Function

r = ds9get([type [, dims],] [access_point,] command; kw...)

Returns the output of an XPA get command sent to DS9.

For complex outputs, one can use the type and optionally the dims parameters, which are directly passed to XPA.get.

DS9ui.ds9set — Function

ds9set([access_point], "command"; kw...)
ds9set([access_point], (; key1=val1, key2=val2...))

Transmits one or more commands to DS9 using the XPA interface.

In the first form, the command is entered as a single string and is transmitted unchanged to DS9. Alternatively, one can pass a NamedTuple with a series of keys and values. Each keywords result in a single command sent to DS9. The value of the keyword is directly used without any interpolation. The only exception is that a Boolean value is converted into "yes"/"no" (following DS9 notation).

DS9ui.ds9cursor — Function

ds9cursor([access_point]; coords=:image, event=:button)

Return the coordinates of the cursor in the DS9 window.

ident: the identifier of the DS9 window.

coords: the type of coordinates to returnL can be :image, :physical, :fk5, :galactic

event: the type of event to capture the cursor position. Can be :button, :key, :any

The function returns a tuple (key, coords), where key is a string representing the event and coords is an array of coordinates as float numbers.

DS9ui.ds9wcs — Function

ds9wcs([access_point]; useheader=true)

Return the WCS transformation of the current image in the DS9 window.

Keyword Arguments

ident: the identifier of the DS9 window.

useheader: if true, the WCS is extracted from the FITS header. If false, the WCS is extracted from the DS9 window.

DS9ui.ds9getregions — Function

ds9getregions([access_point,] name=""; coords=:image, selected=false)

Return the regions defined in the DS9 window.

The optional argument name is the name of the group of regions to extract. If name is an empty string, all regions are extracted.

Keyword Arguments

ident: the identifier of the DS9 window.

coords: the type of coordinates to returnL can be :image, :physical, :fk5, :galactic

The return value is an array of tuples (shape, coordinates, properties), where shape is a symbol indicating shape of the region, coordinates is an array of coordinates, and properties is a dictionary with the properties of the region.